Revert "move demo addresses to test folder"
This reverts commit a8a9953a82d73f8b7e85d045fd36b8fc0ea7172c.
