+++ node/demo_constants.go
@@ -1,4 +1,4 @@
-package demo
+package node
 
 // DeployAccount is the accounts used for development.
 type DeployAccount struct {
